requirements.txt
scipy>=1.4.1
numpy>=1.17.4
pandas>=0.25.1
Pillow>=6.2.1
seaborn>=0.9.0
wordcloud>=1.7.0
matplotlib>=3.1.1
tensorflow>=2.3.0
scikit-image>=0.16.2
scikit-learn>=0.22.1
tensorflow_addons>=0.12.1